Refactor some deps, add docker files, update makefile.
