+++ server/observer/observer.go
@@ -9,15 +9,6 @@ import (
 
 //go:generate mockgen -destination=./mock_conn.go --package=observer net Conn
 
-type ConnHandlerMessage int
-
-const (
-	FILE_LISTENER_CONN_TYPE ConnHandlerMessage = iota
-	FILE_REQUEST_CONN_TYPE
-	SERVER_READY_TO_RECEIVE_FILE_REQUEST
-	SERVER_SENDING_FILE_LIST
-)
-
 type ConnectionObserver struct {
 	Address string
 	Conn    net.Conn
@@ -25,8 +16,8 @@ type ConnectionObserver struct {
 
 func (c *ConnectionObserver) LoadAllFiles(files []common.FileInfo) error {
 
-	if _, err := c.Conn.Write([]byte(fmt.Sprintf("%d\n", SERVER_SENDING_FILE_LIST))); err != nil {
-		fmt.Printf("Unable to write %s to %s\n", SERVER_SENDING_FILE_LIST, c.Address)
+	if _, err := c.Conn.Write([]byte(fmt.Sprintf("%d\n", common.SERVER_SENDING_FILE_LIST))); err != nil {
+		fmt.Printf("Unable to write %s to %s\n", common.SERVER_SENDING_FILE_LIST, c.Address)
 		return fmt.Errorf("error %v\n: ", err)
 	}
 
@@ -55,19 +46,3 @@ func (c *ConnectionObserver) GetIdentifier() string {
 	return c.Address
 }
 
-func IsCHM(n int) bool {
-	conv := ConnHandlerMessage(n)
-	return conv.String() != ""
-}
-
-func (chm ConnHandlerMessage) String() string {
-	switch chm {
-	case FILE_LISTENER_CONN_TYPE:
-		return "FILE_LISTENER_CONNECTION"
-	case FILE_REQUEST_CONN_TYPE:
-		return "FILE_REQUEST_CONNECTION"
-	case SERVER_READY_TO_RECEIVE_FILE_REQUEST:
-		return "SERVER_READY_TO_RECEIVE_FILE_REQUEST"
-	}
-	return ""
-}
